Make "existingfile" work with non-existent default
Implement the `default:` and `env:` tags as resolvers, and change the
`--help` output from the `BeforeApply()` hook to `BeforeResolve()`. This
allows `--help` to work in the presence of a non-existent file as a
default for "existingfile" (and "existingdir"). Previously the defaults
were set very early with `Reset()`, which triggered the existingfile
error for a non-existent file. By moving it to a resolver, they are set
later. By moving the `--help` output from the `BeforeApply()` hook to
the `BeforeResolve()` hook, it becomes possible to see the help message
and what the defaults are before the non-existent file triggers the
error.