Handle flags and args on default commands
Handle flags and args on default commands without having to specify the
command on the command line. That is, a cli spec of

  cmd [subcmd] [--subcmd-flag] <subcmd-arg>

can be executed as

  cmd --subcmd-flag arg

This allows a command to have a primary purpose / subcmd but also
support other subcommands for ancillary tasks, where the primary
subcommand name does not need to be explicitly written out. For cases
where an arg of the default subcommand may be ambiguous with another
subcommand, the other subcommand will take precedence, in which case the
primary subcommand name will need to be specified on the CLI to take
that path.